Deleted test server and other test related classes.
