+++ pom.xml
@@ -116,16 +116,6 @@
                 <artifactId>google-oauth-client</artifactId>
                 <version>1.22.0</version>
             </dependency>
-            <dependency>
-                <groupId>com.google.oauth-client</groupId>
-                <artifactId>google-oauth-client-java6</artifactId>
-                <version>1.20.0</version>
-            </dependency>
-            <dependency>
-                <groupId>com.google.oauth-client</groupId>
-                <artifactId>google-oauth-client-jetty</artifactId>
-                <version>1.20.0</version>
-            </dependency>
 
             <!-- JWT -->
             <dependency>
