+++ proxy-idp/pom.xml
@@ -18,19 +18,16 @@
             <artifactId>proxy-api</artifactId>
         </dependency>
 
-        <!-- https://mvnrepository.com/artifact/com.google.oauth-client/google-oauth-client -->
+        <!-- Google-OAuth-Client -->
         <dependency>
             <groupId>com.google.oauth-client</groupId>
             <artifactId>google-oauth-client</artifactId>
         </dependency>
-        <!-- https://mvnrepository.com/artifact/com.google.api-client/google-api-client -->
+
+        <!-- JWT -->
         <dependency>
-            <groupId>com.google.oauth-client</groupId>
-            <artifactId>google-oauth-client-java6</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.oauth-client</groupId>
-            <artifactId>google-oauth-client-jetty</artifactId>
+            <groupId>com.nimbusds</groupId>
+            <artifactId>nimbus-jose-jwt</artifactId>
         </dependency>
     </dependencies>
