+++ proxy-proxy/pom.xml
@@ -54,24 +54,10 @@
             <artifactId>javax.servlet-api</artifactId>
         </dependency>
 
-        <!-- Google-Oauth-Client -->
+        <!-- Proxy-IDP -->
         <dependency>
-            <groupId>com.google.oauth-client</groupId>
-            <artifactId>google-oauth-client</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.oauth-client</groupId>
-            <artifactId>google-oauth-client-java6</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.oauth-client</groupId>
-            <artifactId>google-oauth-client-jetty</artifactId>
-        </dependency>
-
-        <!-- JWT -->
-        <dependency>
-            <groupId>com.nimbusds</groupId>
-            <artifactId>nimbus-jose-jwt</artifactId>
+            <groupId>no.difi.idporten.oidc</groupId>
+            <artifactId>proxy-idp</artifactId>
         </dependency>
