--- proxy-proxy/src/main/java/httptest/HttpStaticServerInitializer.java
@@ -1,32 +1,0 @@
-package httptest;
-
-/**
- * Created with IntelliJ IDEA.
- * User: I
- * Date: 07.07.13
- * Time: 0:45
- * To change this template use File | Settings | File Templates.
- */
-import io.netty.channel.ChannelInitializer;
-import io.netty.channel.ChannelPipeline;
-import io.netty.channel.socket.SocketChannel;
-import io.netty.handler.codec.http.HttpObjectAggregator;
-import io.netty.handler.codec.http.HttpRequestDecoder;
-import io.netty.handler.codec.http.HttpResponseEncoder;
-import io.netty.handler.stream.ChunkedWriteHandler;
-import io.netty.handler.traffic.ChannelTrafficShapingHandler;
-
-public class HttpStaticServerInitializer extends ChannelInitializer<SocketChannel> {
-    @Override
-    public void initChannel(SocketChannel ch) throws Exception {
-        // Create a default pipeline implementation.
-        ChannelPipeline pipeline = ch.pipeline();
-
-        pipeline.addLast("channelTraffic", new ChannelTrafficShapingHandler(1000));
-        pipeline.addLast("decoder", new HttpRequestDecoder());
-        pipeline.addLast("aggregator", new HttpObjectAggregator(65536));
-        pipeline.addLast("encoder", new HttpResponseEncoder());
-        pipeline.addLast("chunkedWriter", new ChunkedWriteHandler());
-        pipeline.addLast("handler", new HttpStaticServerHandler());
-    }
-}
