+++ proxy-proxy/src/main/java/no/difi/idporten/oidc/proxy/proxy/HttpRequestHandler.java
@@ -2,11 +2,11 @@ package no.difi.idporten.oidc.proxy.proxy;
 
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
-import io.netty.handler.codec.http.HttpHeaderNames;
-import io.netty.handler.codec.http.HttpMessage;
-import io.netty.handler.codec.http.HttpRequest;
+import io.netty.handler.codec.http.*;
+import no.difi.idporten.oidc.proxy.idp.IdportenIdentityProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
 
 /**
  * The first handler that can manipulate incoming HTTP messages and change the remote address we send the request to.
@@ -14,11 +14,18 @@ import org.slf4j.LoggerFactory;
 public class HttpRequestHandler extends SimpleChannelInboundHandler<HttpMessage> {
     private static Logger logger = LoggerFactory.getLogger(HttpRequestHandler.class);
 
+    private final String COOKIE;
+
+    private IdportenIdentityProvider idportenIdentityProvider;
+    private CookieInHeader checkHeaderForCookie;
 
     private final int maxConnectionsQueued;
 
     public HttpRequestHandler(int maxConnectionsQueued) {
         this.maxConnectionsQueued = maxConnectionsQueued;
+        this.COOKIE = "Cookie";
+        idportenIdentityProvider = new IdportenIdentityProvider();
+        checkHeaderForCookie = new CookieInHeader();
     }
 
     @Override
@@ -39,6 +46,8 @@ public class HttpRequestHandler extends SimpleChannelInboundHandler<HttpMessage>
         logger.debug(String.format("HttpRequestHandler reading message: %s", msg));
         if (msg instanceof HttpRequest) {
             HttpRequest request = (HttpRequest) msg;
+
+
             String host = request.headers().getAsString(HttpHeaderNames.HOST);
             logger.debug(String.format("Trimmed host name: %s", host));
         }
